logo-api/requirements.txt
flask
pillow
requests
google-cloud-storage